requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT-Simple==0.0.3
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
PyJWT==1.6.1
six==1.11.0
Werkzeug==0.14.1
zbarlight==2.0
